Server: edit Entity Tag

+++ prolog-spring/src/main/java/com/ssafy/api/service/PostsService.java
@@ -69,13 +69,16 @@ public class PostsService {
                 post.getCoverImage(),post.getCoverColor());
         Post savedPost = postJpaRepo.save(newPost); // post 저장
         //Tag 처리
-        String[] tags = post.getTagList(); // DTO 에서 String[] 으로 태그 받아옴
-        for (String tag :tags) {
-            Tag savedTag = tagJpaRepo.findByTag(tag).orElse(null);
-            if(savedTag == null) {
-                savedTag = tagJpaRepo.save(new Tag(tag));
+        //Tag가 Null일때 -> tag 처리안함
+        if(!post.getTagList().equals(null)){
+            String[] tags = post.getTagList(); // DTO 에서 String[] 으로 태그 받아옴
+            for (String tag :tags) {
+                Tag savedTag = tagJpaRepo.findByTag(tag).orElse(null);
+                if(savedTag == null) {
+                    savedTag = tagJpaRepo.save(new Tag(tag));
+                }
+                tagManageJpaRepo.save(new TagManage(savedPost,savedTag));
             }
-            tagManageJpaRepo.save(new TagManage(savedPost,savedTag));
         }
         return true;
     }
@@ -89,15 +92,18 @@ public class PostsService {
         postOrigin.setUpdate(post.getTitle(),post.getBody(),post.getCoverImage(),post.getCoverColor());
 
         //Tag 처리
-        //기존 태그 Delete
-        tagManageJpaRepo.deleteAllByPost(postOrigin);
-        String[] tags = post.getTagList();
-        for (String tag :tags) {
-            Tag savedTag = tagJpaRepo.findByTag(tag).orElse(null);
-            if(savedTag == null) {
-                savedTag = tagJpaRepo.save(new Tag(tag));
+        //Tag가 Null일때 -> tag 처리안함
+        if(!post.getTagList().equals(null)){
+            //기존 태그 Delete
+            tagManageJpaRepo.deleteAllByPost(postOrigin);
+            String[] tags = post.getTagList();
+            for (String tag :tags) {
+                Tag savedTag = tagJpaRepo.findByTag(tag).orElse(null);
+                if(savedTag == null) {
+                    savedTag = tagJpaRepo.save(new Tag(tag));
+                }
+                tagManageJpaRepo.save(new TagManage(postOrigin,savedTag));
             }
-            tagManageJpaRepo.save(new TagManage(postOrigin,savedTag));
         }
         return postOrigin;
     }
